Split py and py-free methods
