+++ crates/circuit/src/parameter/mod.rs
@@ -1,0 +1,3 @@
+pub mod parameter_expression;
+pub mod symbol_expr;
+pub mod symbol_parser;
